refactor: uses bender's fa2 contract
The bender's FA2 is the same as TQ, without create_token.
The reasoning behind is that we can't expect every FA2 we will manage to have this ep in the token_manager branch (for instance, baseDAO, or NFT).
Managed FA2 will need to be originated with a fixed list of assets.